pyproject.toml
[project]
name = "python-django-module"
version = "0.1.0"
description = ""
authors = [
    {name = "Donato4Humans",email = "donato.4work@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

